refactor(proxy): extract read-failover orchestration into internal/proxy/readpath (#849)
Move the per-key read-failover protocol (broadcastRead, tryEachLocation,
tryBackendsSequentially / InParallel, launchBackendProbes, runBackendProbe,
drainAndCleanupLosers, recordBroadcastWinner, broadcastAllFailed,
failoverFailureResult, withReadFailover, readBackendFn type) and the two
shared sentinels (noopCleanup, errUsageLimitSkip) out of object.Manager
into a new internal/proxy/readpath subpackage.

Shape of the new subpackage:

  type Probe func(...) (size int64, cleanup func(), err error)
  type Failover struct { ... }
  func New(core Core, stores LocationStore, cache LocationCache,
           parallelBroadcast bool) *Failover
  func (f *Failover) Read(ctx, operation, key, probe) (string, error)

Failover owns the span (StartSpan with the existing "Manager <op>" prefix
preserved) and the metric emission, so GET/HEAD callers in object.Manager
hand it a Probe and get back the winning backend name + error. The Probe
type carries the *core.ObjectLocation pointer for non-degraded reads and
nil for degraded broadcasts, matching the historic contract.

object.Manager gains a private *readpath.Failover field built once in
New from the same ObjectCore, MetadataStore, *LocationCache, and
parallelBroadcast flag the Manager already holds. GetObject and
HeadObject become o.failover.Read(...) one-liners (plus the per-call
Probe). objects_read.go shrinks from 1063 to 714 lines.

ObjectCore keeps BackendOrder() with a doc note: no method in the
object package calls it directly, but readpath.Core needs it and
Manager passes its own ObjectCore through to readpath.New, so the
type-check has to be satisfied at this boundary.

broadcast_helpers_test.go (tests for broadcastAllFailed and
drainAndCleanupLosers) moves with the code so the package-private
helpers remain testable.

No behavior change. Full test suite + golangci-lint pass.

Part of #849.
