+++ .version
@@ -1,1 +1,1 @@
-v0.49.0
+v0.50.0
